==> notebooks/1_connections_processing.ipynb <==
# Connections Data Processing
The purpose of this notebook is to process and clean connections data to prepare it for the Neo4j database. 

It will also be used as a reference set for the other datasets in that:
* We only want Station data that has corresponding connection data.
* We only want Iterchange data that has corresponding Station data.

import pandas as pd
import numpy as np
from common_processing import clean_station

df_connections_raw = pd.read_csv(r"../data/raw/connections_raw.csv")
df_connections_raw.head(10)

# Taking the average of the am_peak and inter_peak running times
running_time_av = round((df_connections_raw["AM peak (0700-1000) Running Time (Mins)"] \
    + df_connections_raw["Inter peak (1000 - 1600) Running time (mins)"])/2, 2)

df_connections_raw["running_time_av"] = running_time_av
    
df_connections_raw.head()

# Removing trailing whitespace from Line column (see csv)
df_connections_clean = df_connections_raw[["Line", "Direction", "Station from (A)", "Station to (B)", "Distance (Kms)", "Un-impeded Running Time (Mins)", "running_time_av"]]
df_connections_clean = df_connections_clean.assign(Line=df_connections_raw["Line"].apply(lambda x: x.strip()))

# Correcting Hammersmith & City line
df_connections_clean["Line"] = df_connections_clean["Line"].apply(
    lambda x: "Hammersmith & City" if x == "H & C" else x
)

# Removing "East London" i.e. Overground entries as we won't be including these
df_connections_clean = df_connections_clean[df_connections_clean["Line"] != "East London"]

df_connections_clean.head()

# Common Station processing
df_connections_clean['Station from (A)'] = df_connections_clean['Station from (A)'].apply(clean_station)
df_connections_clean['Station to (B)'] = df_connections_clean['Station to (B)'].apply(clean_station)

# Converting Station names to remove text in 
def clean_connection_station(x):
    idx = x.find(" (")
    station = x if idx == -1 else x[:idx]
    station = "Kings Cross St Pancras" if station == "Kings Cross" else station 
    station = "Walthamstow Central" if station == "Walthamstow" else station
    station = "Highbury & Islington" if station == "Highbury" else station
    station = "Heathrow Terminals 1 2 3" if station == "Heathrow 123" else station
    station = "Heathrow Terminal 4" if station == "Heathrow Terminal Four" else station
    return station

df_connections_clean['Station from (A)'] = df_connections_clean['Station from (A)'].apply(clean_connection_station)
df_connections_clean['Station to (B)'] = df_connections_clean['Station to (B)'].apply(clean_connection_station)

df_connections_clean.head()

df_connections_clean = df_connections_clean.rename(columns={
    "Station from (A)": "station_from",
    "Station to (B)": "station_to",
    "Line": "line",
    "Direction": "direction",
    "Distance (Kms)": "distance_kms",
    "Un-impeded Running Time (Mins)": "un-impeded_running_time"
})
df_connections_clean.head()

df_connections_clean = df_connections_clean[["station_from", "station_to", "line", "direction", "distance_kms", "un-impeded_running_time", "running_time_av"]]
print(len(df_connections_clean))
print(len(df_connections_clean.drop_duplicates()))

# Saving the cleaned dataset
df_connections_clean.to_csv(r"../data/processed/connections_clean.csv")
