==> notebooks/2_stations_processing.ipynb <==
# Station Data Processing
The purpose of this notebook is to clean the Station data set ready for use in the Neo4j database and to ensure it aligns with the data in the Connections dataset.

import pandas as pd
from common_processing import clean_station

df_raw = pd.read_csv(r"../data/raw/stations_raw.csv")
df_raw.head(10)

# Drop duplicate stations duplicate station names that contain the line they are on
stations_to_drop = {
    "Edgware Road (Bakerloo)",
    "Hammersmith (Met)"
}

mask = ~(df_raw["Station"].isin(stations_to_drop))
df_stations = df_raw[mask]

# Common station processing
df_stations.loc[:, "Station"] = df_stations["Station"].apply(clean_station)

# This needs to be renamed
df_stations.loc[df_stations["Station"] == "St Jamess Park", "Station"] = "St James Park"

# Saving the above for review
df_stations.to_csv(r"../data/raw/station_raw_v2.csv")

At this point we now need to remove any stations that aren't present in the Connections dataset

df_connections = pd.read_csv(r"../data/processed/connections_clean.csv", index_col=0)
df_connections.head()

# Checking if that all stations are in one column of the Connections dataset (see the associated notebook for its processing)
df_station_from_a = set(df_connections["station_from"])
df_station_from_b = set(df_connections["station_to"])
print(df_station_from_a.difference(df_station_from_b))
print(df_station_from_b.difference(df_station_from_a))

# Removing stations from the Stations dataset that aren't in the connections dataset
connection_stations = set(df_connections["Station from (A)"])
mask = df_stations['Station'].isin(connection_stations)
df_valid_stations = df_stations[mask].reset_index(drop=True)

# Checking which stations are in Connections but not the cleaned
connection_stations.difference(df_valid_stations["Station"])

df_valid_stations.head()

len(df_valid_stations)

# We now need to add the line that each station is on. There should be duplicate station names where that station is on multiple lines.
# We will need to join to the connections dataset to get this information.
df_stations_with_lines = pd.merge(
    left=df_valid_stations,
    right=df_connections,
    left_on="Station",
    right_on="station_from",
    how="inner"
)
df_stations_with_lines = df_stations_with_lines[["Station", "line", "OS X", "OS Y", "Latitude", "Longitude", "Zone", "Postcode"]].drop_duplicates().reset_index(drop=True)
df_stations_with_lines.head()

len(df_stations_with_lines)

# Renaming columns and saving
df_stations_with_lines.columns = ["name", "line", "os_x", "os_y", "latitude", "longitude", "zone", "postcode"]
df_stations_with_lines.to_csv(r"../data/processed/stations_clean.csv")
