==> notebooks/3_interchange_processing.ipynb <==
# Interchange Data Processing
The purpose of this notebook is to process the interchange data into a format that can be uploaded to the Neo4j database.

import pandas as pd
import numpy as np
from common_processing import clean_station

df_raw = pd.read_csv(r"../data/raw/interchange_raw.csv", header=0, encoding="windows-1252")
df_raw.head(10)

new_col_names = ['station', 'interchange']
df_raw.columns = new_col_names
df_raw.head(5)

df_raw.info()

df_no_nan_rows = df_raw.dropna(axis=0, how='any').reset_index(drop=True)
df_no_nan_rows.head(10)

# Common station processing
df_clean_stations = df_no_nan_rows.assign(
    station = df_no_nan_rows["station"].apply(clean_station)
)

# Additional station processing
df_clean_stations["station"] = df_clean_stations["station"].apply(
    lambda x: x if x.find(",") == -1 else x[:x.find(",")]
)

# Correcting station names
df_clean_stations.loc[df_clean_stations["station"] == "Edgware Road Circle", "station" ] = "Edgware Road"
df_clean_stations.loc[df_clean_stations["station"] == "Hammersmith District & Piccadilly Lines", "station" ] = "Hammersmith"
df_clean_stations.loc[df_clean_stations["station"] == "Kensington Olympia", "station" ] = "Kensington"
df_clean_stations.loc[df_clean_stations["station"] == "Shepherds Bush Market", "station" ] = "Shepherds Bush"
df_clean_stations.loc[df_clean_stations["station"] == "St Jamess Park", "station" ] = "St James Park"
df_clean_stations.loc[df_clean_stations["station"] == "Heathrow Terminal  5", "station" ] = "Heathrow Terminal 5"

# Writing to file so I can review
df_clean_stations.to_csv(r"../data/raw/interchange_raw_v2.csv")

We now need to ensure that the station names align with those stored in the cleaned Station data.

df_stations_ref = (pd.read_csv(r"../data/processed/stations_clean.csv"))
stations_ref = set(df_stations_ref["name"])
interchange_stations = set(df_clean_stations["station"])

stations_to_drop = interchange_stations.difference(stations_ref)
stations_to_drop
# These are all London Underground, DLR or otherwise so can be dropped

mask = ~(df_clean_stations["station"].isin(stations_to_drop))
df_interchange_clean = df_clean_stations[mask].reset_index(drop=True)
df_interchange_clean.head(10)

Bringing in the cleaned Connection dataset and using the lines in it to extract lines from the Interchange data.

df_connections_ref = pd.read_csv(r"../data/processed/connections_clean.csv")
lines_ref = set(df_connections_ref["line"])

def retrieve_lines(interchange, lines):
    """
    This function is used to extract underground lines from the "interchange"
    field of the Interchange dataset. It uses a reference set of underground
    lines to extract underground lines from the text, but only if it identifies
    two underground lines in the text.
    """
    
    lines_and_pos = []
    for line in lines:
        idx = interchange.find(line)
        if idx > -1: lines_and_pos.append((line, idx))
    
    lines_and_pos_sorted = sorted(lines_and_pos, key=lambda item: item[1])
    
    if len(lines_and_pos_sorted) == 1:
        return lines_and_pos_sorted[0][0]
    elif len(lines_and_pos_sorted) == 2:
        return lines_and_pos_sorted[0][0], lines_and_pos_sorted[1][0]
    else:
        return np.nan, np.nan

df_interchange_lines = df_interchange_clean.copy(deep=True)
df_interchange_lines[["line_from", "line_to"]] = df_interchange_clean.loc[:, "interchange"].apply(
    lambda x: retrieve_lines(x, lines_ref)
).apply(pd.Series)

df_interchange_lines.head(10)

def retrieve_duration(x):
    idx_colon = x.find(":")
    idx_mins = x.find(" minutes")
    if (idx_colon == -1) or (idx_mins == -1):
        return np.nan
    else:
        return x[(idx_colon+2):idx_mins]   
    
df_interchange_lines = df_interchange_lines.assign(
    duration = df_interchange_lines['interchange'].apply(retrieve_duration)
)

df_interchange_lines.head(10)

'''
Dropping entires that have only one line identified. Reasons:
    * The other line is not in the reference set.
    * The interchange is between different branches of a line,
        which isn't being considered at this point.
'''
mask = (
    (df_interchange_lines["line_from"].isna() & df_interchange_lines["line_to"].isna()) |
    (df_interchange_lines["line_from"].notna() & df_interchange_lines["line_to"].notna())
)

df_interchange_lines = df_interchange_lines[mask].reset_index(drop=True)

# Writing the data to csv for review
df_interchange_lines.to_csv(r"../data/raw/interchange_raw_v3.csv", encoding="UTF-8")

# Writing the data to another csv that I am going to make maual ammendments to

# The manual ammendments include
#   - Completing interchanges that the text processing wasn't able to process
#   - Removing interchanges between lines that aren't being considered (e.g DLR -> Overground)
#   - Removing interchanges between Bank and Monument and vice versa

# DO NOT RUN AFTER MANUALLY ALTERING FILE
# df_interchange_lines.to_csv(r"../data/raw/interchange_raw_v4.csv", encoding="UTF-8")

df_interchange_v4 = pd.read_csv(r"../data/raw/interchange_raw_v4.csv", index_col=0).drop(columns="interchange").reset_index(drop=True)
df_interchange_v4.head(10)

As the connections dataset is outdated, the interchange dataset contains interchanges between lines at stations where the station may not be on a given line in the connections data set.

As such, we need to use the connections dataset again to remove these interchanges.

print(f"{len(df_interchange_v4)=}")



df_joined_from = pd.merge(
    left=df_interchange_v4,
    right=df_connections_ref[["station_from", "line"]],
    left_on=["station", "line_from"],
    right_on=["station_from", "line"],
    how="inner"
).drop_duplicates().reset_index(drop=True)

df_joined_from_to = pd.merge(
    left=df_joined_from,
    right=df_connections_ref[["station_to", "line"]],
    left_on=["station", "line_to"],
    right_on=["station_to", "line"],
    how="inner"
).drop_duplicates().reset_index(drop=True)

df_joined_from_to = df_joined_from_to[df_interchange_v4.columns]

print(f"{len(df_joined_from)=}")

# Saving to processed folder
df_joined_from_to.to_csv(r"../data/processed/interchanges_clean.csv")
